Invoke new generate method, allow multiple actions
Update the "front-end" of the program so that the user can use the
new "Key Pair Generation" method introduced in the prior commit.

In the long run, the program should be able to use these newly-generated Keys
for different purposes:
(a). for decryption of a secondary Key received from the Server, or at least,
(b). for signing documents to prove identity to the Server.

But in practice, none of this can happen until the user actually manually
contacts the Lead Developer ("Speak To Me" step).

One interesting prototype feature would be to "extract the currently running
JAR, encrypt the 'Payload' class, and re-zip it into a patched Jar file." This
would likely warrant a new class, JarUtils.java, to use the JarInputStream and
whatever else is required to get the currently-running Jar path and then
overwrite it. Stay tuned.

Future Me, never forget that this is a mere prototype! The real version is
expected to use a different, secondary key to run the encryption. And this step
will ideally run under the hood, seamlessly, without the user noticing.

Pretty epic, huh? One day, at long last, I will get developers to Speak To Me.