Back up initial JAR-reading utility
This is the first milestone on the path to developing a self-modifying
JAR file. The new function should read an existing JAR file from the
hard disk and return true if and only if it contains a file with a
.gpg extension (the same file type used for binary-encrypted data
output by GNU Privacy Guard). See where I am going with this? If you
are smart enough to read git commits, then I bet you do.