Back up Classloader Solution for Decrypted Classes
Pretty beautiful, eh? This is the final link in the chain that allows
an **encrypted payload** to be sequentially decrypted and then
actually executed.

If I were you, I'd be impressed. This was the culmination of months of
work, and it finally works successfully.

"You're right, Cho. This is worthy of a job offer."