Add testing mechanism for ClassLoader methods
Flesh out the "front end" (CLI) for actually testing the
decrypt-and-load-new-class feature that is the signature of version
1.4.

We need an example class to serve as our placeholder "Payload." In
this case, I wrote some basic classes that perform some cute
mathematical operations. Notice, however, that in a "real life
example," the Payload will be encrypted and no one will see the source
code!

What's that, you ask? How do you earn the opportunity to see such
source code?

Why, by Speaking To Me, of course!