Add first successful passphrase decryption feature
This is a momentous occasion: for the first time in STM History, we
have managed to develop a Client program that requests a ciphertext
string, reads it correctly via Selenium API calls, parses the ASCII
Armor, and then decrypts the message, all in one fell swoop. The new
feature, "Mode Seven," showcases this feature completely.

In the future, we will be able to use this passphrase _without showing
it to the Client_. This could, I suspect, be massively powerful,
because if the user doesn't know what their current passphrase (aka
"Secondary Key") is, then they'll need to Speak To Me in order to
discover it.

Then, from time to time, the DRM Client will signal the server to
"reset the Secondary Key for this user" and re-encrypt the Payload
with the new one. At the very least, this could help prevent "copying
and redistributing the JAR file," since, if memory serves, it will be
the case that "at any given point in time, only the most recent
Secondary Key will be valid" -- or at least, only the most recent
Secondary Key will be stored in the server's databases!

As always, please do test! Ask questions! Tell me where I'm wrong or
where the program can be improved!

Who am I kidding. You won't do that. You'll make me do it all by
myself.....

.....but not for long. (Mua-ha-ha.)